Separate out logic for deriving packet tag from split filenames
sq packet split generates a directory full of files, and we use the
filenames to derive the packet tag. This was previously done by
checking for specific trailing strings, but this becomes more
error-prone in later versions of `sq`. This refactors for the logic
for sq <= 0.31.0 in preparation for modifying it when supporting sq ==
0.32.0.

Signed-off-by: Ed Smith <ed.smith@collabora.com>