Fix new linter warnings
The annotations in `trust_icon()` indicated it could never be passed
None, which made the linter mark the final return for unknown input as
unreachable. There are tests that pass None, so it's the annotation
that is likely wrong, and has been fixed.

The other changes are harmless whitespace changes made by black.

Signed-off-by: Ed Smith <ed.smith@collabora.com>